experiments/requirements.txt
requests
matplotlib
tqdm
pymongo
python-dotenv
plac
goose3
dateparser
pandas
streamlit
tldextract